Consolidating grain extensions in one place + fixing GetReference() implementation
