+++ Source/Kernel/Orleans/Grains/IGrainType.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Kernel.Grains;
+namespace Orleans;
 
 /// <summary>
 /// Represents a way to specify the grain type.
