+++ Source/Kernel/Orleans/Grains/InvalidGrainName.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Kernel.Grains;
+namespace Orleans;
 
 /// <summary>
 /// Exception that gets thrown when a grain name is invalid for a job.
